+++ apps/ymir-sdl3/src/app/ui/windows/debug/sh2_debugger_window.hpp
@@ -3,6 +3,7 @@
 #include "sh2_window_base.hpp"
 
 #include <app/ui/views/debug/sh2_debug_toolbar_view.hpp>
+#include <app/ui/views/debug/sh2_debugger_model.hpp>
 #include <app/ui/views/debug/sh2_disassembly_view.hpp>
 #include <app/ui/views/debug/sh2_registers_view.hpp>
 
@@ -12,6 +13,8 @@ class SH2DebuggerWindow : public SH2WindowBase {
 public:
     SH2DebuggerWindow(SharedContext &context, bool master);
 
+    void RequestOpen(bool triggeredByEvent, bool requestFocus);
+
     void LoadState(std::filesystem::path path);
     void SaveState(std::filesystem::path path);
 
@@ -20,9 +23,10 @@ protected:
     void DrawContents() override;
 
 private:
+    SH2DebuggerModel m_debuggerModel;
+    SH2DisassemblyView m_disasmView;
     SH2DebugToolbarView m_toolbarView;
     SH2RegistersView m_regsView;
-    SH2DisassemblyView m_disasmView;
 };
 
 } // namespace app::ui